[CLI] Move commands to cmds to avoid module name conflict for now.
